+++ src/core/client.ts
@@ -15,8 +15,17 @@ import {
   runAfterNotifyHandlers
 } from './util'
 import {
-  Config, Logger, BreadcrumbRecord, BeforeNotifyHandler, AfterNotifyHandler, Notice, Noticeable
+  Config,
+  Logger,
+  BeforeNotifyHandler,
+  AfterNotifyHandler,
+  Notice,
+  Noticeable,
+  HoneybadgerStore,
+  BacktraceFrame,
+  BreadcrumbRecord
 } from './types'
+import { DefaultStore } from "./store";
 
 const notifier = {
   name: 'honeybadger-js',
@@ -31,19 +40,11 @@ const TAG_SEPARATOR = /,|\s+/
 const NOT_BLANK = /\S/
 
 export default class Client {
-  /** @internal */
   private __pluginsExecuted = false
 
-  /** @internal */
-  protected __context: Record<string, unknown> = {}
-  /** @internal */
-  protected __breadcrumbs: BreadcrumbRecord[] = []
-  /** @internal */
+  protected __store: HoneybadgerStore<{ context: Record<string, unknown>; breadcrumbs: BreadcrumbRecord[] }> = null;
   protected __beforeNotifyHandlers: BeforeNotifyHandler[] = []
-  /** @internal */
   protected __afterNotifyHandlers: AfterNotifyHandler[] = []
-
-  /** @internal */
   protected __getSourceFileHandler: (path: string, cb: (fileContent: string) => void) => void
 
   config: Config
@@ -75,10 +76,11 @@ export default class Client {
 
       ...opts,
     }
+
     this.logger = logger(this)
   }
 
-  factory(_opts?: Record<string, unknown>): unknown {
+  factory(_opts?: Partial<Config>): Client {
     throw (new Error('Must implement __factory in subclass'))
   }
 
@@ -94,6 +96,9 @@ export default class Client {
       this.__pluginsExecuted = true
       this.config.__plugins.forEach((plugin) => plugin.load(this))
     }
+
+    this.__store = opts.store || new DefaultStore({ context: {}, breadcrumbs: [] })
+
     return this
   }
 
@@ -109,24 +114,31 @@ export default class Client {
 
   setContext(context: Record<string, unknown>): Client {
     if (typeof context === 'object') {
-      this.__context = merge(this.__context, context)
+      const store = this.__store.getStore()
+      store.context = merge(store.context, context)
     }
     return this
   }
 
   resetContext(context?: Record<string, unknown>): Client {
     this.logger.warn('Deprecation warning: `Honeybadger.resetContext()` has been deprecated; please use `Honeybadger.clear()` instead.')
+    const store = this.__store.getStore()
+
     if (typeof context === 'object' && context !== null) {
-      this.__context = merge({}, context)
-    } else {
-      this.__context = {}
+      store.context = context
     }
+    else {
+      store.context = {}
+    }
+
     return this
   }
 
   clear(): Client {
-    this.__context = {}
-    this.__breadcrumbs = []
+    const store = this.__store.getStore()
+    store.context = {}
+    store.breadcrumbs = []
+
     return this
   }
 
@@ -173,11 +185,12 @@ export default class Client {
       }
     })
 
-    notice.__breadcrumbs = this.config.breadcrumbsEnabled ? this.__breadcrumbs.slice() : []
+    const breadcrumbs = this.__store.getStore().breadcrumbs
+    notice.__breadcrumbs = this.config.breadcrumbsEnabled ? breadcrumbs.slice() : []
 
     // we need to have the source file data before the beforeNotifyHandlers,
     // in case they modify them
-    const sourceCodeData = notice ? notice.backtrace.slice() : null
+    const sourceCodeData = notice && notice.backtrace ? notice.backtrace.map(trace => newObject(trace) as BacktraceFrame) : null
 
     getSourceForBacktrace(sourceCodeData, this.__getSourceFileHandler, sourcePerTrace => {
       sourcePerTrace.forEach((source, index) => {
@@ -254,8 +267,9 @@ export default class Client {
       return null
     }
 
+    const context = this.__store.getStore().context
     const noticeTags = this.__constructTags(notice.tags)
-    const contextTags = this.__constructTags(this.__context["tags"])
+    const contextTags = this.__constructTags(context["tags"])
     const configTags = this.__constructTags(this.config.tags)
 
     // Turning into a Set will remove duplicates
@@ -264,7 +278,7 @@ export default class Client {
 
     notice = merge(notice, {
       name: notice.name || 'Error',
-      context: merge(this.__context, notice.context),
+      context: merge(context, notice.context),
       projectRoot: notice.projectRoot || this.config.projectRoot,
       environment: notice.environment || this.config.environment,
       component: notice.component || this.config.component,
@@ -295,7 +309,9 @@ export default class Client {
     const category = opts.category || 'custom'
     const timestamp = new Date().toISOString()
 
-    this.__breadcrumbs.push({
+    const store = this.__store.getStore()
+    let breadcrumbs = store.breadcrumbs
+    breadcrumbs.push({
       category: category as string,
       message: message,
       metadata: metadata as Record<string, unknown>,
@@ -303,25 +319,23 @@ export default class Client {
     })
 
     const limit = this.config.maxBreadcrumbs
-    if (this.__breadcrumbs.length > limit) {
-      this.__breadcrumbs = this.__breadcrumbs.slice(this.__breadcrumbs.length - limit)
+    if (breadcrumbs.length > limit) {
+      breadcrumbs = breadcrumbs.slice(breadcrumbs.length - limit)
     }
+    store.breadcrumbs = breadcrumbs
 
     return this
   }
 
-  /** @internal */
   protected __developmentMode(): boolean {
     if (this.config.reportData === true) { return false }
     return (this.config.environment && this.config.developmentEnvironments.includes(this.config.environment))
   }
 
-  /** @internal */
   protected __send(_notice: Partial<Notice>): void {
     throw (new Error('Must implement send in subclass'))
   }
 
-  /** @internal */
   protected __buildPayload(notice: Notice): Record<string, Record<string, unknown>> {
     const headers = filter(notice.headers, this.config.filters) || {}
     const cgiData = filter({
@@ -362,7 +376,6 @@ export default class Client {
     }
   }
 
-  /** @internal */
   protected __constructTags(tags: unknown): Array<string> {
     if (!tags) {
       return []