+++ src/core/types.ts
@@ -1,4 +1,5 @@
 import Client from './client'
+import { AsyncLocalStorage } from "async_hooks";
 
 export interface Logger {
   log(...args: unknown[]): unknown
@@ -30,6 +31,7 @@ export interface Config {
   filters: string[]
   __plugins: Plugin[],
   tags: unknown,
+  store?: HoneybadgerStore<{ context: Record<string, unknown>; breadcrumbs: BreadcrumbRecord[] }>,
 }
 
 export interface BeforeNotifyHandler {
@@ -94,3 +96,5 @@ export interface CGIData {
   HTTP_COOKIE: string | undefined,
   [x: string]: unknown
 }
+
+export type HoneybadgerStore<T> = Pick<AsyncLocalStorage<T>, 'getStore' | 'run'>