+++ test/unit/server/aws_lambda.server.test.ts
@@ -2,7 +2,7 @@ import Singleton from "../../../src/server";
 // @ts-ignore
 import { nullLogger } from "../helpers";
 import nock from "nock";
-import { spy } from 'sinon'
+import * as sinon from 'sinon'
 import { Notice } from "../../../src/core/types";
 // eslint-disable-next-line import/no-unresolved
 import { APIGatewayProxyEvent, APIGatewayProxyResult, Context } from 'aws-lambda'
@@ -20,6 +20,15 @@ const mockAwsResult = (obj: Partial<APIGatewayProxyResult> = {}) => {
     return Object.assign({}, obj) as APIGatewayProxyResult;
 }
 
+const initNock = (expectedTimes = 1): nock.Scope => {
+    nock.cleanAll()
+
+    return nock("https://api.honeybadger.io")
+        .post("/v1/notices/js")
+        .times(expectedTimes)
+        .reply(201, '{"id":"1a327bf6-e17a-40c1-ad79-404ea1489c7a"}')
+}
+
 describe('Lambda Handler', function () {
     let client: typeof Singleton;
 
@@ -30,13 +39,13 @@ describe('Lambda Handler', function () {
         })
     })
 
-    describe('with arguments', function() {
-        const awsEvent = mockAwsEvent({ body: '1' })
-        const awsContext = mockAwsContext({ awsRequestId: '2' })
+    describe('with arguments', function () {
+        const awsEvent = mockAwsEvent({body: '1'})
+        const awsContext = mockAwsContext({awsRequestId: '2'})
         let handlerFunc;
 
-        beforeEach(function() {
-            handlerFunc = spy()
+        beforeEach(function () {
+            handlerFunc = sinon.spy(() => Promise.resolve())
             const handler = client.lambdaHandler(handlerFunc) as AsyncHandler
             return handler(awsEvent, awsContext)
                 .then(() => {
@@ -48,22 +57,22 @@ describe('Lambda Handler', function () {
                 })
         })
 
-        it('calls original handler with arguments', function() {
+        it('calls original handler with arguments', function () {
             expect(handlerFunc.lastCall.args.length).toBe(2)
             expect(handlerFunc.lastCall.args[0]).toBe(awsEvent)
             expect(handlerFunc.lastCall.args[1]).toBe(awsContext)
         })
     })
 
-    describe('async handlers', function() {
+    describe('async handlers', function () {
 
         it('calls handler with asynchronous response if no error is thrown', async function () {
             client.configure({
                 apiKey: 'testing'
             })
 
-            const handler = client.lambdaHandler(async function(_event, _context) {
-                return Promise.resolve(mockAwsResult({body:'works!'}))
+            const handler = client.lambdaHandler(async function (_event, _context) {
+                return Promise.resolve(mockAwsResult({body: 'works!'}))
             }) as AsyncHandler<APIGatewayProxyEvent, APIGatewayProxyResult>
 
             const res = await handler(mockAwsEvent(), mockAwsContext())
@@ -76,8 +85,8 @@ describe('Lambda Handler', function () {
                 apiKey: 'testing'
             })
 
-            const handler = client.lambdaHandler(async function(_event, _context) {
-                return mockAwsResult({body:'works!'})
+            const handler = client.lambdaHandler(async function (_event, _context) {
+                return mockAwsResult({body: 'works!'})
             }) as AsyncHandler<APIGatewayProxyEvent, APIGatewayProxyResult>
 
             const res = await handler(mockAwsEvent(), mockAwsContext())
@@ -91,33 +100,26 @@ describe('Lambda Handler', function () {
             })
 
             // @ts-expect-error
-            const handler = client.lambdaHandler(async function(_event, _context) {
+            const handler = client.lambdaHandler(async function (_event, _context) {
                 throw new Error("Badgers!")
             }) as AsyncHandler<APIGatewayProxyEvent, APIGatewayProxyResult>
 
             return expect(handler(mockAwsEvent(), mockAwsContext())).rejects.toEqual(new Error("Badgers!"))
         })
 
-        it('reports errors to Honeybadger', async function() {
+        it('reports errors to Honeybadger', async function () {
             client.configure({
                 apiKey: 'testing'
             })
-
-            nock.cleanAll()
-
-            const api = nock("https://api.honeybadger.io")
-                .post("/v1/notices/js")
-                .reply(201, '{"id":"1a327bf6-e17a-40c1-ad79-404ea1489c7a"}')
-
+            const api = initNock()
             // @ts-expect-error
-            const handler = client.lambdaHandler(async function(_event) {
+            const handler = client.lambdaHandler(async function (_event) {
                 throw new Error("Badgers!")
             }) as AsyncHandler<APIGatewayProxyEvent, APIGatewayProxyResult>
 
             try {
                 await handler(mockAwsEvent(), mockAwsContext())
-            }
-            catch (e) {
+            } catch (e) {
                 // eslint-disable-next-line jest/no-conditional-expect
                 expect(e).toEqual(new Error("Badgers!"))
             }
@@ -130,21 +132,14 @@ describe('Lambda Handler', function () {
             })
         })
 
-        // eslint-disable-next-line jest/expect-expect
-        it('reports async errors to Honeybadger', async function() {
+        it('reports async errors to Honeybadger', async function () {
             client.configure({
                 apiKey: 'testing'
             })
-
-            nock.cleanAll()
-
-            const api = nock("https://api.honeybadger.io")
-                .post("/v1/notices/js")
-                .reply(201, '{"id":"1a327bf6-e17a-40c1-ad79-404ea1489c7a"}')
-
-            const handler = client.lambdaHandler(async function(_event) {
+            const api = initNock()
+            const handler = client.lambdaHandler(async function (_event) {
                 return new Promise<APIGatewayProxyResult>((resolve, reject) => {
-                    setTimeout(function() {
+                    setTimeout(function () {
                         reject(new Error("Badgers!"))
                     }, 0)
                 })
@@ -152,8 +147,7 @@ describe('Lambda Handler', function () {
 
             try {
                 await handler(mockAwsEvent(), mockAwsContext())
-            }
-            catch (e) {
+            } catch (e) {
                 // eslint-disable-next-line jest/no-conditional-expect
                 expect(e).toEqual(new Error("Badgers!"))
             }
@@ -167,7 +161,7 @@ describe('Lambda Handler', function () {
         })
     })
 
-    describe('non-async handlers', function() {
+    describe('non-async handlers', function () {
 
         beforeEach(function () {
             client.configure({
@@ -177,8 +171,8 @@ describe('Lambda Handler', function () {
 
         it('calls handler if no error is thrown', function () {
             return new Promise((done) => {
-                const handler = client.lambdaHandler(function(_event, _context, callback) {
-                    callback(null, mockAwsResult({ body: 'works!' }))
+                const handler = client.lambdaHandler(function (_event, _context, callback) {
+                    callback(null, mockAwsResult({body: 'works!'}))
                 }) as SyncHandler<APIGatewayProxyEvent, APIGatewayProxyResult>
 
                 handler(mockAwsEvent(), mockAwsContext(), (err, res) => {
@@ -195,7 +189,7 @@ describe('Lambda Handler', function () {
                     apiKey: null
                 })
 
-                const handler = client.lambdaHandler(function(_event, _context, _callback) {
+                const handler = client.lambdaHandler(function (_event, _context, _callback) {
                     throw new Error("Badgers!")
                 }) as SyncHandler<APIGatewayProxyEvent, APIGatewayProxyResult>
 
@@ -206,14 +200,10 @@ describe('Lambda Handler', function () {
             })
         })
 
-        it('reports errors to Honeybadger', function() {
-            nock.cleanAll()
+        it('reports errors to Honeybadger', function () {
+            const api = initNock()
 
-            const api = nock("https://api.honeybadger.io")
-                .post("/v1/notices/js")
-                .reply(201, '{"id":"1a327bf6-e17a-40c1-ad79-404ea1489c7a"}')
-
-            const handler = client.lambdaHandler(function(_event, _context, _callback) {
+            const handler = client.lambdaHandler(function (_event, _context, _callback) {
                 throw new Error("Badgers!")
             }) as SyncHandler<APIGatewayProxyEvent, APIGatewayProxyResult>
 
@@ -228,15 +218,11 @@ describe('Lambda Handler', function () {
             })
         })
 
-        it('reports async errors to Honeybadger', function() {
-            nock.cleanAll()
+        it('reports async errors to Honeybadger', function () {
+            const api = initNock()
 
-            const api = nock("https://api.honeybadger.io")
-                .post("/v1/notices/js")
-                .reply(201, '{"id":"1a327bf6-e17a-40c1-ad79-404ea1489c7a"}')
-
-            const handler = client.lambdaHandler(function(_event, _context, callback) {
-                setTimeout(function() {
+            const handler = client.lambdaHandler(function (_event, _context, callback) {
+                setTimeout(function () {
                     callback(new Error("Badgers!"))
                 }, 0)
             }) as SyncHandler<APIGatewayProxyEvent, APIGatewayProxyResult>
@@ -253,25 +239,21 @@ describe('Lambda Handler', function () {
         })
 
         it('calls beforeNotify and afterNotify handlers', function () {
-            nock.cleanAll()
+            const api = initNock()
 
-            const api = nock("https://api.honeybadger.io")
-                .post("/v1/notices/js")
-                .reply(201, '{"id":"1a327bf6-e17a-40c1-ad79-404ea1489c7a"}')
-
-            const handler = client.lambdaHandler(function(_event, _context, callback) {
-                setTimeout(function() {
+            const handler = client.lambdaHandler(function (_event, _context, callback) {
+                setTimeout(function () {
                     callback(new Error("Badgers!"))
                 }, 0)
             }) as SyncHandler<APIGatewayProxyEvent, APIGatewayProxyResult>
 
             client.beforeNotify(function (notice: Notice) {
-                notice.context = Object.assign(notice.context, { foo: 'bar' })
+                notice.context = Object.assign(notice.context, {foo: 'bar'})
             })
 
             let afterNotifyCalled = false;
             client.afterNotify(function (err: Error | undefined, notice: Notice) {
-                expect(notice.context).toEqual({ foo: 'bar' })
+                expect(notice.context).toEqual({foo: 'bar'})
                 afterNotifyCalled = true;
             })