+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased][beta]
+### Added
+- AsyncLocalStorage for AWS Lambda Handler (#688)
+
 ### Fixed
 - Allow special characters in tags. Also support space-delimited tags:
   "one two three" and "one, two, three" are equivalent